+++ src/engine/bvh.h
@@ -34,7 +34,7 @@ struct BVHNode {
 
 std::vector<BVHNode> BuildBVHTree(std::vector<MeshObject> objects);
 
-std::vector<int> FrustumCull(const std::vector<BVHNode>& nodes,
+std::vector<size_t> FrustumCull(const std::vector<BVHNode>& nodes,
                              const base::Frustumf& frustum);
 
 void DumpBVHTree(const std::vector<BVHNode>& nodes,
